Updating the connect options to these three: 1. Connecting with the server name, username, and password 2. Connecting with a connection string 3. Connecting with a ConnectionConfig object, which is used to generate a connection string.
The main difference is the ConnectionConfig struct. This replaces the
old ConnectionInfo struct, which unnecessarily stored attributes in a
hash map before generating a connection string. The ConnectionInfo
struct, in addition to having unnecessary allocations and hashing, was
also clunkier to use - it required a few extra steps such as
initializing the ConnectionInfo, setting fields, and calling a
getConnectionString function.